+++ nav2_bringup/bringup/maps/urjc_restaurant.yaml
@@ -1,0 +1,7 @@
+image: urjc_restaurant.pgm
+resolution: 0.020000
+origin: [-25.960000, -27.240000, 0.000000]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.196
+
